requirements.txt
pyspark
pandas

apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-mysql
apache-airflow-providers-openlineage>=1.8.0

